Experimental OM2 flipped face fix
After spending probably far too much time on trying to find something in the format that declares 'THESE FACES ARE BACKWARDS' I've decided to just implement a hack as it seems there's nothing in the file format for this. Either KFIV didn't use backface culling, or I'm interpreting the file very badly.